Failing test case for interaction with uuid2 and String uniqueidentifier
Test requires a MS SQL Server (tested with version 11.0.2100.60).
The generated id with uuid2 is lower case. If that id is used to
retrieve the entity from the database, the entity is fetched as well as
the element collection, but due to SQL Server returning uniqueidentifier
columns as upper case, the owner for the collection cannot be found.
The culprit is the equals check of EntityKey between the entity (lower
case id) and the collection (upper case id) when attempting to retrieve
the owner via StatefulPersistenceContext#getEntity(EntityKey).